Move sginfo and add license file.
