+++ src/data/CMakeLists.txt
@@ -28,4 +28,4 @@ include_directories(data PRIVATE ${PROJECT_SOURCE_DIR}/src)
 target_link_libraries(data PRIVATE base)
 
 add_subdirectory(ff)
-add_subdirectory(sg)
+add_subdirectory(sginfo)
