+++ src/data/sginfo/sgclib.c
