+++ src/data/sginfo/sgsi.c
